# contracts/api/openapi/vocal_coach_v1.yaml
openapi: 3.1.0
info:
  title: Vocal Coach API
  version: 1.0.0
paths:
...
  /v1/ai/jobs:
    get:
      summary: List user AI evaluation jobs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
  /v1/ai/jobs/{job_id}:
    get:
      summary: Get AI evaluation job details
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  /v1/sessions:
    get:
      summary: List current-user sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
    post:
      summary: Create session
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
...
  /v1/sessions/{session_id}/feedback:
    get:
      summary: Get session feedback
      security:
        - bearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  /v1/karaoke/catalog:
    get:
      summary: Get full karaoke drill catalog
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Karaoke catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KaraokeCatalogOut'
        "401":
          description: Unauthorized
...
  /v1/karaoke/catalog/{drill_id}:
    get:
      summary: Get single karaoke drill detail
      security:
        - bearerAuth: []
      parameters:
        - name: drill_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Drill detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KaraokeDrillOut'
        "401":
          description: Unauthorized
        "404":
          description: Drill not found
  /v1/karaoke/progress:
    get:
      summary: Get current-user karaoke progress
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
  /v1/karaoke/evaluate-tone:
    post:
      summary: Evaluate a karaoke audio sample
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio, pitch_score, rhythm_delay_ms]
              properties:
                audio:
                  type: string
                  format: binary
                pitch_score:
                  type: number
                  minimum: 0
                  maximum: 100
                rhythm_delay_ms:
                  type: number
                  minimum: -5000
                  maximum: 5000
      responses:
        "200":
          description: Tone evaluation
          content:
            application/json:
              schema:
                type: object
                required: [pitch_score, rhythm_delay_ms, tone_score, ai_feedback, feedback_source]
                properties:
                  pitch_score:
                    type: number
                    minimum: 0
                    maximum: 100
                  rhythm_delay_ms:
                    type: number
                    minimum: -5000
                    maximum: 5000
                  tone_score:
                    type: [number, "null"]
                    minimum: 0
                    maximum: 100
                  ai_feedback:
                    type: string
                  feedback_source:
                    type: string
                    enum: [local_audio_analysis, pitch_timing_only]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    metricsToken:
      type: apiKey
      in: header
      name: X-Metrics-Token
  schemas:
...
    MelodyNote:
      type: object
      required: [note, start_beat, duration_beats]
      properties:
        note:
          type: string
          minLength: 1
        start_beat:
          type: number
          minimum: 0
        duration_beats:
          type: number
          exclusiveMinimum: 0
    KaraokeDrillOut:
      type: object
      required:
        - drill_id
        - title
        - style_category
        - difficulty
        - duration_sec
        - tempo_bpm
        - vocal_range
        - objective
        - performance_tips
        - melody_reference
      properties:
        drill_id:
          type: string
        title:
          type: string
        style_category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        duration_sec:
          type: integer
          minimum: 10
          maximum: 600
        tempo_bpm:
          type: integer
          minimum: 40
          maximum: 220
        vocal_range:
          type: object
          additionalProperties:
            type: string
        objective:
          type: string
        performance_tips:
          type: array
          minItems: 1
          items:
            type: string
        melody_reference:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/MelodyNote'
        instrumental_url:
          type: string
        pitch_map_url:
          type: string
        artist_name:
          type: string
        cover_url:
          type: string
    KaraokeCategoryOut:
      type: object
      required: [category_id, style_label, description, drills]
      properties:
        category_id:
          type: string
        style_label:
          type: string
        description:
          type: string
        drills:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/KaraokeDrillOut'
    KaraokeCatalogOut:
      type: object
      required: [module_id, title, description, categories]
      properties:
        module_id:
          type: string
        title:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/KaraokeCategoryOut'
...